backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosmtplib==3.0.1
email-validator==2.2.0
slowapi==0.1.9
httpx==0.27.0
